Perform minor refactorings and style/text changes to welcome banner
