+++ frontend/src/app/app.component.spec.ts
@@ -8,8 +8,8 @@ import { RouterTestingModule } from '@angular/router/testing'
 import { async, TestBed } from '@angular/core/testing'
 import { AppComponent } from './app.component'
 import { NavbarComponent } from './navbar/navbar.component'
-import { WelcomeComponent } from './welcome-banner/welcome.component'
-import { ChallengeSolvedNotificationComponent } from 'src/app/challenge-solved-notification/challenge-solved-notification.component'
+import { WelcomeComponent } from './welcome/welcome.component'
+import { ChallengeSolvedNotificationComponent } from './challenge-solved-notification/challenge-solved-notification.component'
 
 import { MatSelectModule } from '@angular/material/select'
 import { MatFormFieldModule } from '@angular/material/form-field'
