+++ frontend/src/app/welcome-banner/welcome-banner.component.html
@@ -1,4 +1,4 @@
-<div class="mat-body" >  
+<div class="mat-typography" >
     <h1><span translate [translateParams]="{appname: applicationName}">WELCOME_BANNER_TITLE</span></h1>
     <div class="text-justify" [innerHtml]="'WELCOME_BANNER_DESCRIPTION' | translate:{appname: applicationName}"></div>
     <button mat-raised-button class="close-dialog welcome-banner-close-button" (click)="closeWelcome()"><i class="fas fa-arrow-circle-left fa-lg"></i>
