+++ frontend/src/assets/i18n/en.json
@@ -190,6 +190,6 @@
   "GDPR_MESSAGE": "We take data security, customer privacy, and legal compliance very serious. In accordance with GDPR we allow you to request complete erasure of your account and any associated data.",
   "DELETE_DATA_LABEL": "Delete User Data",
   "INFO_HACKING_INSTRUCTOR": "Start an interactive hacking tutorial for this challenge.",
-  "WELCOME_BANNER_TITLE":"Welcome to the {{appname}}",
-  "WELCOME_BANNER_DESCRIPTION": "<p>Being a web application with a vast number of intended security vulnerabilities, the {{appname}} is supposed to be the opposite of a best practice or template application for web developers: It is an awareness, training, demonstration and exercise tool for security risks in modern web applications. The {{appname}} is an open-source project hosted by the non-profit Open Web Application Security Project (OWASP) and is developed and maintained by volunteers.</p><p><a href=\"https://leanpub.com/juice-shop\">Pwning OWASP Juice Shop Book</a><br /><a href=\"https://github.com/bkimminich/juice-shop\">GitHub: https://github.com/bkimminich/juice-shop</a></p>"
+  "WELCOME_BANNER_TITLE":"Welcome to the {{appname}}!",
+  "WELCOME_BANNER_DESCRIPTION": "<p>Being a web application with a vast number of intended security vulnerabilities, the <strong>{{appname}}</strong> is supposed to be the opposite of a best practice or template application for web developers: It is an awareness, training, demonstration and exercise tool for security risks in modern web applications. The <strong>{{appname}}</strong> is an open-source project hosted by the non-profit <a href='https://owasp.org' target='_blank'>Open Web Application Security Project (OWASP)</a> and is developed and maintained by volunteers. Check out the link below for more information and documentation on the project.</p><h1><a href='http://owasp-juice.shop' target='_blank'>http://owasp-juice.shop</a></h1>"
 }
